frontend/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
openai>=1.81.0
langchain-upstage>=0.1.0
langgraph>=0.1.0
langchain-core>=0.2.0
pydantic>=2.0.0
typing-extensions>=4.8.0 